backend/pyproject.toml
[project]
name = "rootstock-api"
version = "0.1.0"
description = "Rootstock management API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "gitpython>=3.1,<4",
    "ruamel.yaml>=0.18,<1",
    "httpx>=0.28,<1",
    "cryptography>=44,<45",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "ruff>=0.8,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
